+++ templates/Virtual-Assistant-Template/typescript/generator-botbuilder-assistant/generators/app/templates/customAssistant/test/mocha.opts
@@ -2,4 +2,3 @@
 --timeout 50000
 --recursive
 --colors
-./**/*Test.js
