+++ templates/Virtual-Assistant-Template/typescript/generator-botbuilder-assistant/generators/app/templates/customAssistant/test/mocks/nockFixtures/interruption_confirm_cancel_response.json
