+++ templates/Virtual-Assistant-Template/typescript/generator-botbuilder-assistant/generators/app/templates/customAssistant/test/mocks/nockFixtures/mainDialog_greeting_response.json
