+++ templates/Virtual-Assistant-Template/typescript/generator-botbuilder-assistant/generators/app/templates/customAssistant/test/mocks/nockFixtures/onboardingDialog_validation.json
